requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain-groq
langgraph
langgraph-cli[inmem]
youtube_transcript_api
nltk
streamlit